requirements.txt
python-dotenv
pytest
google-generativeai
srt
openai-whisper>=20240930
torch>2.4.0
PyYAML
requests